back/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.10
Flask==3.1.0
Flask-Cors==5.0.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-speech==2.29.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.11
nltk==3.9.1
numpy==2.0.2
packaging==24.2
preshed==3.0.9
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
thinc==8.3.3
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.0
zipp==3.21.0
